+++ R/gllvmTMB.R
@@ -703,6 +703,7 @@ gllvmTMB <- function(
       unit_internal  = site,
       family         = family,
       weights        = weights,
+      REML           = REML,
       call           = match.call(),
       is_y_observed  = observed_response$is_y_observed
     ))